+++ src/frontend.nim
@@ -2,9 +2,20 @@ import std/dom
 
 include pkg/karax/prelude
 
+import seamlesserverpkg/renderer/base
 import seamlesserverpkg/renderer/state
 import seamlesserverpkg/routes
 import seamlesserverpkg/renderer/routes/default/notFound
+
+var brData: BridgedData
+
+proc render(rendered: Rendered): VNode =
+  ## Renders the rendered object
+  brData = rendered.parseBridgedData
+  document.title = rendered.genTitle brData.appName
+  result = buildHtml(tdiv):
+    rendered.renderErrors
+    rendered.vnode
 
 proc main =
   var s = newState()
@@ -12,19 +23,18 @@ proc main =
   proc renderer(data: RouterData): VNode =
     buildHtml(tdiv):
       var noRoute = true
-      for (path, renderHtml) in allRoutes:
+      for (path, renderHtmlPage) in allRoutes:
         if $window.location.pathname == path:
-          renderHtml s
+          render renderHtmlPage s
           noRoute = false
           break
       if noRoute:
-        notFound.renderHtml s
+        render notFound.renderHtml s
   setRenderer renderer
 
-  document.addEventListener("pushstate", proc(ev: Event) =
+  document.addEventListener("pushstate") do (ev: Event):
     echo "push state"
     preventDefault ev 
-  )
 
 when isMainModule:
   main()