+++ src/seamlesserverpkg/renderer/base.nim
@@ -1,7 +1,33 @@
+from std/strformat import fmt
+
 when defined js:
   import std/dom
 
 include pkg/karax/prelude
+
+type
+  Rendered* = ref object
+    vnode*: VNode
+    title*: string
+    errors*: seq[string]
+  BridgedData* = ref object
+    appName*: string
+
+const brDataDomJsonId = kstring "bridgedData"
+
+using
+  rendered: Rendered
+
+proc addError*(rendered; msg: string) =
+  ## Add new error
+  rendered.errors.add msg
+
+proc renderErrors*(rendered): VNode =
+  ## Renders all erros into a `div`
+  buildHtml(tdiv(class = "errors")):
+    for error in rendered.errors:
+      tdiv(class = "error"):
+        text error
 
 proc dynamicLink*(href = ""; title = ""): VNode =
   ## Creates a default HTML link in backend and in frontend it just pushes the
@@ -11,24 +37,55 @@ proc dynamicLink*(href = ""; title = ""): VNode =
       proc onclick(ev: Event; n: VNode) =
         preventDefault ev
         window.history.pushState("", "", href)
-        echo "pushed"
+
+func genTitle*(rendered; appName: string): string =
+  ## Gets the rendered page title
+  fmt"{rendered.title} - {appName}"
 
 when not defined js:
   from std/os import `/`
+  from std/json import `$`, `%*`, `%`
 
   import ../config
 
-  proc baseHtml*(bodyNode: VNode): string =
+  func genBridgedData(appName: string): VNode =
+    ## Generates the data that'll be sent to Javascript backend
+    let node = $ %*{
+      "appName": appName
+    }
+    buildHtml(script(id = brDataDomJsonId, `type` = "application/json")):
+      verbatim node
+  proc inbaseHtml*(rendered): string =
+    ## Creates the full HTML page with rendered page
     withConf:
       let
         appName = appName
         jsDir = jsDir
     let vnode = buildHtml(html):
       head:
-        title: text appName
-        meta(name="viewport", content="width=device-width, initial-scale=1.0")
+        title: text rendered.genTitle appName
+        meta(name="viewport", content="width=device-width, -scale=1.0")
       body:
         tdiv(id = "ROOT"):
-          bodyNode
+          rendered.renderErrors
+          rendered.vnode
+        genBridgedData(
+          appName = appName
+        )
         script(src = "/" & jsDir / "main.js")
     result = $vnode
+else:
+  from std/json import parseJson, to
+
+  proc parseBridgedData*(rendered): BridgedData =
+    ## Parses the BridgedData sent by server in a script tag
+    new result
+    try:
+      result = document.
+        getElementById(brDataDomJsonId).
+        innerText.
+        `$`.
+        parseJson.
+        to BridgedData
+    except:
+      rendered.addError getCurrentExceptionMsg()