+++ src/seamlesserverpkg/renderer/routes/default/notFound.nim
@@ -1,17 +1,18 @@
 include pkg/karax/prelude
 
 import ../../state
+import ../../base
 
-proc renderHtml*(state: State): VNode =
+proc renderHtml*(state = newState()): Rendered =
   ## Renders the HTML for homepage
-  buildHtml(tdiv):
+  new result
+  result.title = "Not Found"
+  result.vnode = buildHtml(tdiv):
     h1: text "404: Not found"
 
 when not defined js:
   import pkg/prologue
-  
-  import ../../base
 
   proc render*(ctx: Context) {.async.} =
     ## Server side homepage renderer
-    resp baseHtml renderHtml newState()
+    resp renderHtml().inBaseHtml