+++ src/seamlesserverpkg/renderer/routes/user/login.nim
@@ -3,9 +3,11 @@ include pkg/karax/prelude
 import ../../state
 import ../../base
 
-proc renderHtml*(state: State): VNode =
+proc renderHtml*(state = newState()): Rendered =
   ## Renders the HTML for homepage
-  buildHtml(tdiv):
+  new result
+  result.title = "Login"
+  result.vnode = buildHtml(tdiv):
     h1: text "Login page"
     dynamicLink(href = "/"):
       text "Homepage"
@@ -15,4 +17,4 @@ when not defined js:
   
   proc render*(ctx: Context) {.async.} =
     ## Server side homepage renderer
-    resp baseHtml renderHtml newState()
+    resp renderHtml().inBaseHtml